code with old remains

+++ host/result/killjob.sh
@@ -1,4 +1,12 @@
 process="jobinfo.txt"
-proID=`grep 'Submitted batch job' $process |awk '{print $4}'`
-scancel $proID
+RESULTDIR="$( pwd -P )"
+if  [ -f $RESULTDIR/$process ]; then
+  proID=`grep 'Submitted batch job' $process |awk '{print $4}'`
+  if [ -z $proID ]; then
+    echo "Job already finished"
+  else
+    scancel $proID
+  fi
+fi
+
